+++ ws-server/index.ts
@@ -1,0 +1,5 @@
+import { main } from './src';
+
+const store: unknown[] = [];
+
+main(store);
