+++ ws-server/src/prepare-server/index.ts
@@ -1,0 +1,1 @@
+export { prepareServer } from './prepare-server';
