+++ ws-server/src/prepare-web-socket/index.ts
@@ -1,0 +1,1 @@
+export { prepareWebSocket } from './prepare-web-socket';
